Added the cuda part of the code

+++ exam1/omp/floyd_omp.cu
@@ -3,11 +3,11 @@
 #include "floyd_omp.h"
 #include "utils.h"
 
-void floyd_omp(uint8_t node_count, uint8_t *** adj_matrix)
+void floyd_omp(int node_count, int *** adj_matrix)
 {
-    uint8_t k = 0;
-    uint8_t i = 0;
-    uint8_t j = 0;
+    int k = 0;
+    int i = 0;
+    int j = 0;
 
 #   pragma omp parallel for num_threads(omp_get_num_procs()) default(none) shared(adj_matrix, node_count) private(i, k, j)
     for(k = 0; k < node_count; k++)
