+++ exam1/serial/floyd_s.h
@@ -2,6 +2,6 @@
 #define FLOYD_S_H
 #include <stdint.h>
 
-void floyd_serial(uint8_t node_count, uint8_t *** adj_matrix);
+void floyd_serial(int node_count, int *** adj_matrix);
 
 #endif
