+++ exam1/utils.h
@@ -6,9 +6,9 @@
 #include <stdbool.h>
 #include <string.h>
 
-uint8_t minimum(uint8_t left, uint8_t right);
-void copy_adj(const uint8_t size, uint8_t ** adj, uint8_t *** copy); 
-bool check_answer(uint8_t size, uint8_t ** key, uint8_t ** attempt);
-void print_matrix(uint8_t size, uint8_t ** matrix);
+int minimum(int left, int right);
+void copy_adj(const int size, int ** adj, int *** copy); 
+bool check_answer(int size, int ** key, int ** attempt);
+void print_matrix(int size, int ** matrix);
 
 #endif
